chore: move events logger to deployments subpackage
This change moves the deployment events logger to an events package
under the deployments package. It used to be in the openapi package but
the exact functionality is needed to log events when processing Gram
functions so it was lifted out to a more appropriate package.
