+++ server/internal/deployments/events/log.go
@@ -1,4 +1,4 @@
-package openapi
+package events
 
 import (
 	"context"
